fix docs and structure
Translate docs into English
Split into 2 files
